Externalizing contexts from App.tsx
