+++ package.json
@@ -25,6 +25,8 @@
     "@expo/metro-config": "^0.7.1",
     "@expo/metro-runtime": "^1.1.1",
     "@react-native-async-storage/async-storage": "1.17.11",
+    "@react-navigation/bottom-tabs": "^6.5.8",
+    "@react-navigation/drawer": "^6.6.3",
     "@react-navigation/native": "6.1.2",
     "@react-navigation/stack": "^6.3.16",
     "expo": "48.0.6",
